Move default identifier out of io.l5d and update kind name.
