+++ linkerd/protocol/http/src/main/resources/META-INF/services/io.buoyant.linkerd.IdentifierInitializer
@@ -1,1 +1,1 @@
-io.l5d.identifier.DefaultHttpIdentifierInitializer
+io.buoyant.linkerd.protocol.DefaultHttpIdentifierInitializer
