+++ linkerd/protocol/http/src/main/scala/io/buoyant/linkerd/protocol/HttpIdentifierConfig.scala
@@ -1,9 +1,10 @@
-package io.l5d
+package io.buoyant.linkerd.protocol
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility
 import com.fasterxml.jackson.annotation.{JsonAutoDetect, JsonIgnore, JsonTypeInfo}
-import com.twitter.finagle.{Dtab, Path, http}
-import io.buoyant.router.RoutingFactory.Identifier
+import com.twitter.finagle.http.Request
+import com.twitter.finagle.{Dtab, Path}
+import io.buoyant.router.RoutingFactory._
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "kind")
 @JsonAutoDetect(fieldVisibility = Visibility.ANY)
@@ -14,6 +15,5 @@ trait HttpIdentifierConfig {
   def newIdentifier(
     prefix: Path,
     baseDtab: () => Dtab = () => Dtab.base
-  ): Identifier[http.Request]
+  ): Identifier[Request]
 }
-
